gmod_map_list.py: Don't leave a bunch of FDs open.
Probably not a problem as you'd need over 1000 mounted VPKs/addons but
may as well prevent an issue.  The downside to this is some "legacy"
mods are lzma compressed and after reading their header in, then opening
the file again to read the data, some amount of retreading will need to
be done over the LZMA compressed data, but the header isn't very big so
this shouldn't amount to much.